[Add]  -add some define to get rid of build warnings. [Change]  -use uint32_t instead of typedef uint  -change WildlandsEngine to .lib [Fix]  -bug1:  -reason:

+++ Wildlands/src/Wildlands/Platforms/Windows/WindowsInput.cpp
@@ -26,7 +26,7 @@ namespace Wildlands
 		GLFWwindow* window = static_cast<GLFWwindow*>(Application::Get().GetWindow().GetNativeWindow());
 		double xpos, ypos;
 		glfwGetCursorPos(window, &xpos, &ypos);
-		return std::pair<float, float>(xpos, ypos);
+		return std::pair<float, float>(static_cast<float>(xpos), static_cast<float>(ypos));
 	}
 	float WindowsInput::GetMouseXInternal()
 	{
